LPS-52768 moves java
