+++ modules/apps/asset/asset-publisher-web/src/com/liferay/asset/publisher/web/AssetPublisherPortletLayoutListener.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portlet.assetpublisher;
+package com.liferay.asset.publisher.web;
 
+import com.liferay.asset.publisher.web.util.AssetPublisherUtil;
 import com.liferay.portal.kernel.portlet.PortletLayoutListener;
 import com.liferay.portal.kernel.portlet.PortletLayoutListenerException;
 import com.liferay.portal.kernel.util.StringPool;
@@ -24,7 +25,6 @@ import com.liferay.portal.model.PortletPreferences;
 import com.liferay.portal.service.LayoutLocalServiceUtil;
 import com.liferay.portal.service.SubscriptionLocalServiceUtil;
 import com.liferay.portlet.asset.util.AssetUtil;
-import com.liferay.portlet.assetpublisher.util.AssetPublisherUtil;
 import com.liferay.portlet.journal.service.JournalArticleLocalServiceUtil;
 
 /**
