+++ modules/apps/asset/asset-publisher-web/src/com/liferay/asset/publisher/web/portlet/email/dependencies/email_asset_entry_added_body.tmpl
