+++ modules/apps/asset/asset-publisher-web/src/com/liferay/asset/publisher/web/portlet/route/asset-publisher-friendly-url-routes.xml
