+++ modules/apps/asset/asset-publisher-web/src/com/liferay/asset/publisher/web/portlet/template/dependencies/portlet_display_template_rich_summary.ftl
