+++ modules/apps/asset/asset-publisher-web/src/com/liferay/asset/publisher/web/util/CustomUserAttributesAssetEntryQueryProcessor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.assetpublisher.util;
+package com.liferay.asset.publisher.web.util;
 
 import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
 import com.liferay.portal.kernel.util.GetterUtil;
